refactor geo convert service to its own class
